move defaults from actors/ to consts in src/
